requirements.txt
aiohttp
aiosignal
async-timeout
asynctest
attrs
auditee @ file:///home/azureuser/mbehnia/scratch/auditee
base58
bcolors
bitarray
blessings
certifi
cffi
charset-normalizer
click
colorama
cryptography
cytoolz
ecdsa
eth-abi
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing
eth-utils
frozenlist
gitdb
GitPython
hexbytes
idna
importlib-metadata
importlib-resources
ipfshttpclient
jsonschema
lru-dict
multiaddr
multidict
netaddr
parsimonious
pkgutil_resolve_name
protobuf
pycparser
pycryptodome
pydantic
pyrsistent
python-on-whales
PyYAML
requests
rlp
six
smmap
toolz
tqdm
typeguard
typer
typing_extensions
urllib3
varint
web3
websockets
yarl
zipp
